+++ rust/fizzbuzz/benches/bench_fizzbuzz.rs
@@ -1,6 +1,8 @@
 #![allow(dead_code)]
+use std::borrow::Cow;
+
 use criterion::{criterion_group, criterion_main, Criterion};
-use fizzbuzz::{self, FizzBuzz, MultiFizzBuzz};
+use fizzbuzz::{self, FizzBuzz, FizzBuzzAnswer, MultiFizzBuzz};
 use rayon::prelude::*;
 
 static TEST_SIZE: isize = 1_000_000;
@@ -43,27 +45,27 @@ fn vec_pariter() {
 }
 
 #[inline]
-fn multifizzbuzz_trait() {
+fn multifizzbuzz_trait_as_vec_string() {
     let inputs: Vec<_> = (1..TEST_SIZE).collect();
-    let _: Vec<String> = inputs.fizzbuzz().into();
+    let _: Vec<String> = inputs.fizzbuzz().collect();
 }
 
 #[inline]
-fn multifizzbuzz_trait_as_string() {
+fn multifizzbuzz_trait_as_vec_cow() {
     let inputs: Vec<_> = (1..TEST_SIZE).collect();
-    let _: String = inputs.fizzbuzz().into();
+    let _: Vec<Cow<str>> = inputs.fizzbuzz().collect();
 }
 
 #[inline]
 fn multifizzbuzz_trait_from_vec_as_answer() {
     let inputs: Vec<_> = (1..TEST_SIZE).collect();
-    let _ = inputs.fizzbuzz();
+    let _: Vec<FizzBuzzAnswer> = inputs.fizzbuzz().collect();
 }
 
 #[inline]
 fn multifizzbuzz_trait_from_range_as_answer() {
     let inputs = 1..TEST_SIZE;
-    let _ = inputs.fizzbuzz();
+    let _: Vec<FizzBuzzAnswer> = inputs.fizzbuzz().collect();
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
@@ -74,9 +76,11 @@ fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("vec_iter", |b| b.iter(|| vec_iter()));
     // c.bench_function("vec_intoiter", |b| b.iter(|| vec_intoiter()));
     c.bench_function("vec_pariter", |b| b.iter(|| vec_pariter()));
-    c.bench_function("multifizzbuzz_trait", |b| b.iter(|| multifizzbuzz_trait()));
-    c.bench_function("multifizzbuzz_trait_as_string", |b| {
-        b.iter(|| multifizzbuzz_trait_as_string())
+    c.bench_function("multifizzbuzz_trait_as_vec_string", |b| {
+        b.iter(|| multifizzbuzz_trait_as_vec_string())
+    });
+    c.bench_function("multifizzbuzz_trait_as_vec_cow", |b| {
+        b.iter(|| multifizzbuzz_trait_as_vec_cow())
     });
     c.bench_function("multifizzbuzz_trait_from_vec_as_answer", |b| {
         b.iter(|| multifizzbuzz_trait_from_vec_as_answer())